[S] Staggered commits and calendar changes.
I did a massive amount of changes to the project without committing to the repo so go me. Therefore I will be breaking the commit down into smaller chunks. Some code might not make sense to begin with but wait until the end commit (Marked by [F]) and all should 'work'.

In this change I have simply switched the google autoloader from a deprecated version into the correct one (the one I was using before was just a redirect by the looks of it).