Moving errors.
The title is a bit misleading... I have moved a method into a general Utils class so every function can access it without lots of copy-paste-ing. This allows all classes to read a file and respond to errors correctly when reading.
I have also added more error checking to some parts of the system but there is still a lot more to be added.